// src/app/core/_services/constants.enum.ts
export enum Constants {
   
    CONTENT_TYPE = "Content-Type",
    ACCEPT = "Accept",
    moduleSelected = "moduleSelected",
    token = "token",
    FIRST_NAME = "firstName",
    LAST_NAME = "lastName",
    MOBILE = "mobile",
    SIGN_UP_DATA = "signUpData",
    GRADE = "grade",
    GRADEORDER = "gradeorder",
    USERTYPE = "usertype",
    GRADEID = "gradeId",
    MEDIUMID = "mediumId",
    MEDIUMNAME = "mediumName",
    OTP = "otp",
    PASSWORD = "password",
    ID = "_id",
    USER_DETAILS = "UserDetails",
    SUBJECT_ID = "subjectId",
    SOCIALDATA = 'socialData',
    SUBSCRIPTIONCODEISUSED = 'subscriptioncodeused',
    SUBSCRIPTIONCODEVALIDFROM = 'subscriptionCodeValidFrom',
    SUBSCRIPTIONCODEVALIDTO = 'subscriptionCodeValidTo',
    SUBSCRIPTIONCODETITLE = 'subscriptionCodeTitle',
    SUBSCRIPTIONCODECODE = 'subscriptionCodeCode',
    SUBSCRIPTIONCODEACTUALCODE = 'subscriptionCodeActualCode',
    SUBSCRIPTIONCODESECONDLEFT = 'subscriptionCodeSecondleft',
    LASTSUBSCRIPTIONMINUTE = 'lastsubscriptionminute',
    SplashImage = 'splashImage',
    SplashVideo = 'SplashVideo',
    SplashVideoSkipped = 'SplashVideoSkipped',
    AppName = 'Marathi DG Madhyam',
    PAYMENTMETHOD = 'paymentmethod',
    TRANSACTIONID = 'transactionID',
    UNIQUEID = 'uniqueid',
    ISUSERLOGIN = "isUserlogin",
    CITY = 'city',
    DISTRICT = 'district',
    ADDRESS = 'address',
    PINCODE = 'pincode',
    ACTIONTYPE = 'actiontype',
    NOTIFICATIONCOUNT = 'notificationcount',
    DATACONTENTTYPE = 'contentType',
    LOGINTYPE = 'loginType',
    ADDRESSFORMSKIPCLICK = 'addressformskipclick',
    TESTTMELEFT = 'testtimeleft',
    SCHOOLVERIFIED = 'schoolVerified',
    SCHOOLID = 'schoolId',
    DIVISIONID = "divisionId",
    PHOTO = "photo",
    EMAIL = "email",
    BOARD = "board",
    COMPETITIONATTEMPTEDQUESTINOS= "competition-attempted-questions",
    //ZOOMEMAIL = "zoomEmail",
}